add service and repo for project

+++ src/main/java/org/blank/projectmanagementsystem/repository/SystemOutlineRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface SystemOutlineRepository extends JpaRepository<SystemOutline, Long>{
+public interface SystemOutlineRepository extends JpaRepository<SystemOutline, Integer>{
     SystemOutline findByName(String systemOutlineName);
 }
